LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-focus-2_test3-expected.txt

Button1  Button2  Invoker0  Invoker1  Button3  Button4

FAIL Popover focus navigation with imperative invocation assert_equals: Focus should move to popover0 since it is focusable expected Element node <div popover="" id="popover0" tabindex="0" style="top:300... but got Element node <button id="invoker1" tabindex="0">Invoker1</button>
